aggiornato layout contatore-bottom su tutti i template per ottimizzazione layout e aggiunto layout paginazione su tutte le liste
