+++ src/main/java/com/freelapp/controller/ClientController.java
@@ -48,6 +48,12 @@ import jakarta.validation.Valid;
 @Controller
 @ControllerAdvice //  serve per la gestione degli errori di maxSize file
 public class ClientController {
+	
+	//variabile che memorizza l'ultima pagina consultata nella lista Clienti serve per mantenerla durante la sessione
+	private int currentPageListaClienti = 1;
+	
+	//variabile che passo al model del search clienti per dirgli che siamo in modalità search
+	private boolean searchMode = false;
 
 	@Autowired
 	private ClienteRepository repositoryCliente;
@@ -78,6 +84,10 @@ public class ClientController {
 	@GetMapping("/Clienti")
 	public String listaClienti(Model model) {
 		
+		//essendo fuori dalla modalità search reinizializzo la varibile
+		searchMode = false;
+		model.addAttribute("searchMode", searchMode);
+		
 		//passo al model i contatore e task in uso (gli static)
 		model.addAttribute("contatoreInUso", ContatoreController.contatoreInUso);
 		model.addAttribute("taskInUso", ContatoreController.taskInUso);
@@ -102,12 +112,19 @@ public class ClientController {
 		}
 		model.addAttribute("areClientsOnDb", areClientsOnDb);
 		
-		return getOnePage(1, model );
+		return getOnePage(currentPageListaClienti, model );
 	} 
 	
 	
 	@GetMapping("/Clienti/page/{pageNumber}")
 	public String getOnePage(@PathVariable("pageNumber") int currentPage, Model model ) {
+		
+		//aggiorna la variabile che memorizza l'ultima pagina visitata
+		currentPageListaClienti = currentPage;
+					
+		//essendo fuori dalla modalità search reinizializzo la varibile
+		searchMode = false;
+		model.addAttribute("searchMode", searchMode);
 		
 		Page<Cliente> page = clienteService.findPage(currentPage);
 		
@@ -170,6 +187,10 @@ public class ClientController {
 	@GetMapping("/cliente-search")
 	public String listaClientiSearch(@Param("input") String input,Model model) {
 		
+		//passo al model questo booleano per dirgli che siamo in modalità search
+		searchMode = true;
+		model.addAttribute("searchMode", searchMode);
+		
 		//passo al model i contatore e task in uso (gli static)
 		model.addAttribute("contatoreInUso", ContatoreController.contatoreInUso);
 		model.addAttribute("taskInUso", ContatoreController.taskInUso);
@@ -197,67 +218,76 @@ public class ClientController {
 		return clienteBySearch(1, input, model);
 	} 
 		 
-	 @GetMapping("/cliente-search/page/{numberPage}")
-	 public String clienteBySearch(@PathVariable("pageNumber") int currentPage, String input,
+	 @GetMapping("/cliente-search-input={input}/page/{numberPage}")
+	 public String clienteBySearch(@PathVariable("numberPage") int currentPage, @PathVariable("input") String input,
 			 	Model model) {
 				 
-				 Page<Cliente> page = clienteService.findSearchedPage(currentPage,input);
+			// passo al model l'input inserito per mostrare all'utente cosa ha inserito come
+			// input
+			model.addAttribute("inputInserito", input);
+			// passo al model questo booleano per dirgli che siamo in modalità search
+			searchMode = true;
+			model.addAttribute("searchMode", searchMode);
 
-				 int totalPages = page.getTotalPages();
-				 
-				 long totalItems = page.getTotalElements();
-				 
-				 List<Cliente> listaClientiSearch = page.getContent();
-				
-				model.addAttribute("currentPage", currentPage);
-			
-				model.addAttribute("totalPages", totalPages);
-				
-				model.addAttribute("totalItems", totalItems);
-				
-				model.addAttribute("list", listaClientiSearch);	
-				
-				contatoreservice.importContatoreInGet(model);
-				
-				//passo al model l'endpoint da dare come input hidden a start/pause/stop del contatore
-				if(input != null) {
-					String endPoint = "/cliente-search?input=" + input;
-					
-					model.addAttribute("endPoint", endPoint);						
-				} else {
-					String endPoint = "/cliente-search";
-					model.addAttribute("endPoint", endPoint);	
-				}
-				
-				//passo al model i contatore e task in uso (gli static)
-				model.addAttribute("contatoreInUso", ContatoreController.contatoreInUso);
-				model.addAttribute("taskInUso", ContatoreController.taskInUso);
-				
-				//invio al model il booleano del contatore attivato
-				//se contatoreAttivato = true avvio animazione su titolo task al contatore;
-				model.addAttribute("contatoreAttivato", ContatoreController.contatoreAttivato);
-		
-				//inizializzo a false così che al refresh o cambio pagina non esegue animazione ma solo allo start
-				ContatoreController.contatoreAttivato = false;
-				
-				//metodo che passa al model le informazioni sul task in uso per generare la modale STOP
-				taskService.informationFromTaskInUsoToModel(model);
-				
-				//passa al model la lista di tutti i task esclusi quelli chiusi
-				List<Task> taskList = new ArrayList<Task> ();
-				taskList = repositTask.findAllNotClosed();
-				model.addAttribute("taskList", taskList);
-				
-				//restituisce al model questo valore booleano false se non ci sono clienti a db
-				//e restituisce true se ci sono clienti a db
-				boolean areClientsOnDb = false;
-				if(!repositoryCliente.findAll().isEmpty()) {
-					areClientsOnDb = true;
-				}
-				model.addAttribute("areClientsOnDb", areClientsOnDb);
-				
-				return "Clienti/freelApp-listClient";
-		 
+			Page<Cliente> page = clienteService.findSearchedPage(currentPage, input);
+
+			int totalPages = page.getTotalPages();
+
+			long totalItems = page.getTotalElements();
+
+			List<Cliente> listaClientiSearch = page.getContent();
+
+			model.addAttribute("currentPage", currentPage);
+
+			model.addAttribute("totalPages", totalPages);
+
+			model.addAttribute("totalItems", totalItems);
+
+			model.addAttribute("list", listaClientiSearch);
+
+			contatoreservice.importContatoreInGet(model);
+
+			// passo al model l'endpoint da dare come input hidden a start/pause/stop del
+			// contatore
+			if (input != null) {
+				String endPoint = "/cliente-search?input=" + input;
+
+				model.addAttribute("endPoint", endPoint);
+			} else {
+				String endPoint = "/cliente-search";
+				model.addAttribute("endPoint", endPoint);
+			}
+
+			// passo al model i contatore e task in uso (gli static)
+			model.addAttribute("contatoreInUso", ContatoreController.contatoreInUso);
+			model.addAttribute("taskInUso", ContatoreController.taskInUso);
+
+			// invio al model il booleano del contatore attivato
+			// se contatoreAttivato = true avvio animazione su titolo task al contatore;
+			model.addAttribute("contatoreAttivato", ContatoreController.contatoreAttivato);
+
+			// inizializzo a false così che al refresh o cambio pagina non esegue animazione
+			// ma solo allo start
+			ContatoreController.contatoreAttivato = false;
+
+			// metodo che passa al model le informazioni sul task in uso per generare la
+			// modale STOP
+			taskService.informationFromTaskInUsoToModel(model);
+
+			// passa al model la lista di tutti i task esclusi quelli chiusi
+			List<Task> taskList = new ArrayList<Task>();
+			taskList = repositTask.findAllNotClosed();
+			model.addAttribute("taskList", taskList);
+
+			// restituisce al model questo valore booleano false se non ci sono clienti a db
+			// e restituisce true se ci sono clienti a db
+			boolean areClientsOnDb = false;
+			if (!repositoryCliente.findAll().isEmpty()) {
+				areClientsOnDb = true;
+			}
+			model.addAttribute("areClientsOnDb", areClientsOnDb);
+
+			return "Clienti/freelApp-listClient";
 		 
 	  }
