+++ src/main/resources/templates/Progetti/freelapp-descrizioneProgetto.html
@@ -16,7 +16,10 @@
 					extend: {
 						colors: {
 							clifford: '#da373d',
-						}
+						},
+						screens: {
+							xs: '375px'
+						},
 					}
 				}
 			}
@@ -38,7 +41,7 @@
 
 	<body class="font-sans"
 		style="width: 100vw; height: 100vh;" onload="timerstart()">
-		<section id="rapidButton" class=" fixed md:top-[80%] top-[65%] right-0">
+		<section id="rapidButton" class=" fixed bottom-[20%] xs:bottom-[15%]  right-0">
 			<button
 				class=" pe-4 bg-[#FFE541] border-2 border-[#0057A5] rounded-l-lg p-2 text-[#0057A5] font-bold shadow-md shadow-[#0057A5] ">
 				<div class="t-tooltip col-auto w-8 h-8 ms-2">
@@ -524,74 +527,90 @@
 							</div>
 						</section>
 					</div>
-				<div class="pb-28 md:pb-2 lg:pt-2 pt-4 w-11/12 lg:hidden" id="dettaglio-contatore-bottom">
-					<div class="w-full h-fit border-4 border-[#0057A5] bg-[#ffe541] rounded-[35px] p-1">
-						<div class="flex justify-between items-center text-[#0057A5] font-bold">
-							<div id="icona-contatore" class="col-auto ps-3 w-[15%]">
-								<img class="h-[45px] w-[45px]" src="/img/sources/icons/contatore-blue.svg" alt="contatore">
+				<!-- CONTATORE -->
+				<div class="m-auto pb-44 md:pb-2 lg:pt-2 pt-4 w-11/12 lg:hidden"
+					id="dettaglio-contatore-bottom">
+					<div
+						class="w-full h-fit border-4 border-[#0057A5] bg-[#ffe541] rounded-[35px] p-1">
+						<div
+							class="flex justify-between items-center text-[#0057A5] font-bold">
+							<div id="icona-contatore" class="max-xs:hidden col-auto ps-3 w-[15%]">
+								<img class="h-[45px] w-[45px]"
+									src="/img/sources/icons/contatore-blue.svg" alt="contatore">
 							</div>
-							<div th:if="${taskInUso == null}" class="flex flex-col items-center">
-								<div th:unless="${contatoreIsTrue}" class="col-auto text-[14px] p-0">...</div>
+							<div th:if="${taskInUso == null}"
+								class="flex flex-col items-center justify-center w-auto max-xs:ps-4">
+								<div th:unless="${contatoreIsTrue}"
+									class="col-auto text-[14px] p-0">...</div>
 								<div th:if="${contatoreIsTrue}" th:text="${taskInUso.getName()}"
 									class="col-auto text-[14px] p-0"></div>
 								<div class="col-auto text-[24px] p-0" id="timerDue">00:00:00</div>
 							</div>
-							<a th:unless="${taskInUso == null}" th:href="@{|/Task/${taskInUso.id}|}"
-								class="t-tooltip flex flex-col items-center relative w-[45%]">
-								<div th:unless="${contatoreIsTrue}" class="col-auto text-[14px] p-0">...</div>
+							<a th:unless="${taskInUso == null}"
+								th:href="@{|/Task/${taskInUso.id}|}"
+								class="t-tooltip flex flex-col items-center relative w-[45%] justify-center w-auto max-xs:ps-4">
+								<div th:unless="${contatoreIsTrue}"
+									class="col-auto text-[14px] p-0">...</div>
 								<div th:if="${contatoreIsTrue}" th:text="${taskInUso.getName()}"
 									class="col-auto text-[14px] text-left p-0 w-full ml-4 truncate"></div>
 								<div class="col-auto text-[24px] p-0" id="timerDue">00:00:00</div>
 								<div
 									class="t-box hidden text-nowrap border border-[#0057A5] bg-gray-100 rounded-lg absolute p-1 ms-[-210px] mt-[40px]">
-									Dettaglio task
-								</div>
+									Dettaglio task</div>
 							</a>
-							<div class="col-auto flex items-center w-[40%]">
+							<div class="col-auto flex items-center w-[40%] max-xs:w-full max-xs:ps-2">
 								<div th:if="${taskInUso == null}" class="mx-2 flex items-center">
-									<img class="h-[45px] w-[45px]" src="/img/sources/icons/stop-blue.svg" alt="stop">
+									<img class="h-[45px] w-[45px]"
+										src="/img/sources/icons/stop-blue.svg" alt="stop">
 								</div>
-								<div th:unless="${taskInUso == null}" class="mx-2 flex items-center">
-									<button id="stopButtonBottom" type="button" class="hover:opacity-75"
-											onclick="creaModaleStop()">
-										<img class="h-[45px] w-[45px]" src="/img/sources/icons/stop-blue.svg"
-											alt="stop">
+								<div th:unless="${taskInUso == null}"
+									class="mx-2 flex items-center">
+									<button id="stopButtonBottom" type="button"
+										class="hover:opacity-75">
+										<img class="h-[45px] w-[45px]"
+											src="/img/sources/icons/stop-blue.svg" alt="stop">
 									</button>
 								</div>
 								<div th:if="${taskInUso == null}" class="mx-2 flex items-center">
-									<img class="h-[45px] w-[45px]" src="/img/sources/icons/pause-blue.svg" alt="pause">
+									<img class="h-[45px] w-[45px]"
+										src="/img/sources/icons/pause-blue.svg" alt="pause">
 								</div>
-								<div th:unless="${taskInUso == null}" class="mx-2 flex items-center">
+								<div th:unless="${taskInUso == null}"
+									class="mx-2 flex items-center">
 									<div th:unless="${contatoreIsRun}">
-										<img class="h-[45px] w-[45px]" src="/img/sources/icons/pause-blue.svg"
-											alt="pause">
+										<img class="h-[45px] w-[45px]"
+											src="/img/sources/icons/pause-blue.svg" alt="pause">
 									</div>
 									<div th:if="${contatoreIsRun}">
-										<form th:action="@{|/Contatore/pause/${taskInUso.id}|}" th:object="${contatore}"
-											method="post">
+										<form th:action="@{|/Contatore/pause/${taskInUso.id}|}"
+											th:object="${contatore}" method="post">
 											<input type="hidden" name="endPoint" th:value="${endPoint}">
-											<button type="submit" class="hover:opacity-75 flex items-center">
-												<img class="h-[45px] w-[45px]" src="/img/sources/icons/pause-blue.svg"
-													alt="pause">
+											<button type="submit"
+												class="hover:opacity-75 flex items-center">
+												<img class="h-[45px] w-[45px]"
+													src="/img/sources/icons/pause-blue.svg" alt="pause">
 											</button>
 										</form>
 									</div>
 								</div>
 								<div th:if="${taskInUso == null}" class="mx-2 flex items-center">
-									<img class="h-[45px] w-[45px]" src="/img/sources/icons/play-blue.svg" alt="start">
+									<img class="h-[45px] w-[45px]"
+										src="/img/sources/icons/play-blue.svg" alt="start">
 								</div>
-								<div th:unless="${taskInUso == null}" class="mx-2 flex items-center">
+								<div th:unless="${taskInUso == null}"
+									class="mx-2 flex items-center">
 									<div th:if="${contatoreIsRun}">
-										<img class="h-[45px] w-[45px]" src="/img/sources/icons/play-blue.svg"
-											alt="start">
+										<img class="h-[45px] w-[45px]"
+											src="/img/sources/icons/play-blue.svg" alt="start">
 									</div>
 									<div th:unless="${contatoreIsRun}">
-										<form th:action="@{|/start/${taskInUso.id}|}" th:object="${contatore}"
-											method="post">
+										<form th:action="@{|/start/${taskInUso.id}|}"
+											th:object="${contatore}" method="post">
 											<input type="hidden" name="endPoint" th:value="${endPoint}">
-											<button type="submit" class="hover:opacity-75 flex items-center">
-												<img class="h-[45px] w-[45px]" src="/img/sources/icons/play-blue.svg"
-													alt="start">
+											<button type="submit"
+												class="hover:opacity-75 flex items-center">
+												<img class="h-[45px] w-[45px]"
+													src="/img/sources/icons/play-blue.svg" alt="start">
 											</button>
 										</form>
 									</div>
@@ -600,6 +619,7 @@
 						</div>
 					</div>
 				</div>
+				<!-- FINE CONTATORE -->
 				</div>
 		</main>
