+++ src/main/resources/templates/Task/freelapp-descrizioneTask.html
@@ -15,7 +15,10 @@
 					extend: {
 						colors: {
 							clifford: '#da373d',
-						}
+						},
+						screens: {
+							xs: '375px'
+						},
 					}
 				}
 			}
@@ -37,7 +40,7 @@
 
 <body class="font-sans" style="width: 100vw; height: 100vh;"
 	onload="timerstart()">
-	<section id="rapidButton" class=" fixed md:top-[80%] top-[65%] right-0">
+	<section id="rapidButton" class="fixed bottom-[20%] xs:bottom-[15%]  right-0">
 		<button
 			class=" pe-4 bg-[#FFE541] border-2 border-[#0057A5] rounded-l-lg p-2 text-[#0057A5] font-bold shadow-md shadow-[#0057A5] ">
 			<div class="t-tooltip col-auto w-8 h-8 ms-2">
@@ -391,21 +394,19 @@
 				class="gap-4 rounded-lg bg-gray-100 p-6 m-3 shadow-lg">
 				</div>
 			</div>
-
-
-			
-				<div class="pb-28 md:pb-2 lg:pt-2 pt-4 w-11/12 lg:hidden"
+<!-- CONTATORE -->
+				<div class="m-auto pb-44 md:pb-2 lg:pt-2 pt-4 w-11/12 lg:hidden"
 					id="dettaglio-contatore-bottom">
 					<div
 						class="w-full h-fit border-4 border-[#0057A5] bg-[#ffe541] rounded-[35px] p-1">
 						<div
 							class="flex justify-between items-center text-[#0057A5] font-bold">
-							<div id="icona-contatore" class="col-auto ps-3 w-[15%]">
+							<div id="icona-contatore" class="max-xs:hidden col-auto ps-3 w-[15%]">
 								<img class="h-[45px] w-[45px]"
 									src="/img/sources/icons/contatore-blue.svg" alt="contatore">
 							</div>
 							<div th:if="${taskInUso == null}"
-								class="flex flex-col items-center">
+								class="flex flex-col items-center justify-center w-auto max-xs:ps-4">
 								<div th:unless="${contatoreIsTrue}"
 									class="col-auto text-[14px] p-0">...</div>
 								<div th:if="${contatoreIsTrue}" th:text="${taskInUso.getName()}"
@@ -414,7 +415,7 @@
 							</div>
 							<a th:unless="${taskInUso == null}"
 								th:href="@{|/Task/${taskInUso.id}|}"
-								class="t-tooltip flex flex-col items-center relative w-[45%]">
+								class="t-tooltip flex flex-col items-center relative w-[45%] justify-center w-auto max-xs:ps-4">
 								<div th:unless="${contatoreIsTrue}"
 									class="col-auto text-[14px] p-0">...</div>
 								<div th:if="${contatoreIsTrue}" th:text="${taskInUso.getName()}"
@@ -424,7 +425,7 @@
 									class="t-box hidden text-nowrap border border-[#0057A5] bg-gray-100 rounded-lg absolute p-1 ms-[-210px] mt-[40px]">
 									Dettaglio task</div>
 							</a>
-							<div class="col-auto flex items-center w-[40%]">
+							<div class="col-auto flex items-center w-[40%] max-xs:w-full max-xs:ps-2">
 								<div th:if="${taskInUso == null}" class="mx-2 flex items-center">
 									<img class="h-[45px] w-[45px]"
 										src="/img/sources/icons/stop-blue.svg" alt="stop">
@@ -432,7 +433,7 @@
 								<div th:unless="${taskInUso == null}"
 									class="mx-2 flex items-center">
 									<button id="stopButtonBottom" type="button"
-										class="hover:opacity-75" onclick="creaModaleStop()">
+										class="hover:opacity-75">
 										<img class="h-[45px] w-[45px]"
 											src="/img/sources/icons/stop-blue.svg" alt="stop">
 									</button>
